added more tests + some params models changed

+++ app/schemas/Variant.py
@@ -1,13 +1,14 @@
 from pydantic import BaseModel
 from sqlalchemy import Text
 from typing import Optional, List
+from decimal import Decimal
 
 class Variant(BaseModel):
    id: Optional[int] = None
    sku: str
-   sales_price: float
+   sales_price: Decimal
    product_id: Optional[int] = None
-   purchase_price: float
+   purchase_price: Decimal
    type: str
    config_attributes: Optional[List[dict]]
